Solucionado error decimales al añadir zapatilla

+++ src/main/java/MenuAdministrador.java
@@ -55,7 +55,7 @@ public class MenuAdministrador {
         System.out.print("Número: ");
         int numero = Integer.parseInt(teclado.nextLine());
         System.out.print("Precio: ");
-        float precio = Float.parseFloat(teclado.nextLine());
+        double precio = Double.parseDouble(teclado.nextLine());
         Zapatilla zapatilla = new Zapatilla(nombre.trim(), color.trim(), numero, precio);
         catalogoZapatillas.add(zapatilla);
     }
